+++ frontend/src/components/main/timer/timer.js
@@ -1,7 +1,8 @@
 const timerForm = () => {
     const timer = `
     <div class="circle">
-        <svg width="225" viewBox="9 7 230 235" xmlns="http://www.w3.org/2000/svg">
+        <object type="image/svg+xml" data="./potion.svg" class="circle-svg"></object>
+        <svg width="240" viewBox="5 6 230 230" xmlns="http://www.w3.org/2000/svg">
             <g transform="translate(110,110)">
                 <circle r="100" class="e-c-base"/>
                 <g transform="rotate(-90)">
