+++ frontend/src/pages/noticeDetailPage.js
@@ -17,9 +17,9 @@ export default class noticeDetailPage extends AbstractView {
                 <section class="big-content-container">
                         <div class="notice-title-wrapper">이것은 제목입니다.</div>
                         <div class="notice-content-wrapper">이것은 내용입니다.</div>
-                        <button class="notice-edit-btn">수정</button>
-                        <button class="notice-delete-btn">삭제</button>
                         <button class="link-to-notice">목록</button>
+                        <button class="notice-delete-btn"> 삭제</button>
+                        <button class="notice-edit-btn">수정</button>
                 </section>
             </div>
         </div>
