+++ frontend/src/styles/attend.scss
@@ -37,7 +37,18 @@
     box-sizing: border-box;
     padding: 2.5vh 5vw;
 }
-
+.main-content-container {
+    position: absolute;
+    top: 20vh;
+    left: 10vw;
+    width: 80vw;
+    height: 70vh;
+    background-color: #fff;
+    border-radius: 25px;
+    box-shadow: 0 10px 10px rgba(0, 0, 0, 0.5);
+    box-sizing: border-box;
+    padding: 0.5vh 2vw;
+}
 .calendar-header {
     display: flex;
     flex-direction: row;
