+++ frontend/src/styles/timer.scss
@@ -4,12 +4,6 @@
     border-radius: 98px;
     transition: 1s;
 }
-.circle {
-    backdrop-filter: blur(8px);
-    height: 200px;
-    width: 200px;
-    border-radius: 100px;
-}
 
 .pink {
     stroke: #f7958e;
